backend/fix: fix type mismatch for a `addUTCTime` call in `findBookingIdBySpecialZoneOTP` function
* The call to `addUTCTime` in function `findBookingIdBySpecialZoneOTP` had a
  type mis-match in it's first argument.

* addUTCTime needed a first arg of type NominalDiffTime, while the arg provided
  to it was a hard Int type.

* Adds a fromIntegral call on the Int to help convert it into NominalDiffTime

+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Storage/Queries/Booking.hs
@@ -96,7 +96,7 @@ findBookingBySpecialZoneOTP merchantId otpCode now = do
 
 findBookingIdBySpecialZoneOTP :: MonadFlow m => Id Merchant -> Text -> UTCTime -> Int -> m (Maybe (Id Booking))
 findBookingIdBySpecialZoneOTP (Id merchantId) otpCode now bookingOtpExpiry = do
-  let otpExpiryCondition = addUTCTime (- (bookingOtpExpiry * 60) :: NominalDiffTime) now
+  let otpExpiryCondition = addUTCTime (fromIntegral (- (bookingOtpExpiry * 60)) :: NominalDiffTime) now
   (Domain.Types.Booking.id <$>) <$> findOneWithKV [Se.And [Se.Is BeamB.specialZoneOtpCode $ Se.Eq (Just otpCode), Se.Is BeamB.providerId $ Se.Eq merchantId, Se.Is BeamB.createdAt $ Se.GreaterThanOrEq otpExpiryCondition, Se.Is BeamB.status $ Se.Eq NEW]]
 
 cancelBookings :: MonadFlow m => [Id Booking] -> UTCTime -> m ()
